Address PETR example review feedback
