+++ LICENSE
@@ -224,6 +224,7 @@ the following copyright holders, licensed under the Apache License, Version 2.0
   Copyright (c) 2024 Heming Xia
   Copyright 2025 The Qwen team, Alibaba Group and the HuggingFace Inc. team
   Copyright (c) OpenMMLab. All rights reserved.
+  Copyright (c) 2022 megvii-model. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not
 use these files except in compliance with the License. You may obtain a copy
