+++ examples/onnx_ptq/far3d/evaluate.py
@@ -21,9 +21,10 @@
 import argparse
 import importlib
 import os
+import sys
 import warnings
+from pathlib import Path
 
-import tensorrt as trt
 import torch
 from mmcv import Config, DictAction
 from mmcv.utils import import_modules_from_strings
@@ -33,115 +34,9 @@ from mmdet3d.datasets import build_dataset
 from projects.mmdet3d_plugin.datasets.builder import build_dataloader
 from tqdm import tqdm
 
-TRT_TO_TORCH = {
-    trt.DataType.FLOAT: torch.float32,
-    trt.DataType.HALF: torch.float16,
-    trt.DataType.INT8: torch.int8,
-    trt.DataType.INT32: torch.int32,
-    trt.DataType.BOOL: torch.bool,
-    trt.DataType.UINT8: torch.uint8,
-}
-if int(trt.__version__.split(".")[0]) >= 10:
-    TRT_TO_TORCH[trt.DataType.INT64] = torch.int64
+sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
 
-TRT_LOGGER = trt.Logger(trt.Logger.WARNING)
-trt.init_libnvinfer_plugins(TRT_LOGGER, "")
-
-
-def aligned_tensor(shape, dtype, device, alignment=256):
-    element_size = torch.empty((), dtype=dtype).element_size()
-    element_count = int(torch.tensor(shape).prod().item())
-    storage = torch.empty(element_count + alignment // element_size, dtype=dtype, device=device)
-    offset_bytes = (-storage.data_ptr()) % alignment
-    offset = offset_bytes // element_size
-    return storage[offset : offset + element_count].view(shape)
-
-
-class TensorRTRunner:
-    def __init__(self, engine_path, state_names=()):
-        with open(engine_path, "rb") as engine_file:
-            engine_bytes = engine_file.read()
-        self.engine = trt.Runtime(TRT_LOGGER).deserialize_cuda_engine(engine_bytes)
-        if self.engine is None:
-            raise RuntimeError(f"Failed to deserialize {engine_path}")
-        self.context = self.engine.create_execution_context()
-        if self.context is None:
-            raise RuntimeError(f"Failed to create an execution context for {engine_path}")
-        self.tensor_names = [
-            self.engine.get_tensor_name(index) for index in range(self.engine.num_io_tensors)
-        ]
-        self.input_shapes = {}
-        self.output_shapes = {}
-        self.tensor_dtypes = {}
-        for name in self.tensor_names:
-            shape = tuple(self.engine.get_tensor_shape(name))
-            dtype = TRT_TO_TORCH[self.engine.get_tensor_dtype(name)]
-            self.tensor_dtypes[name] = dtype
-            if self.engine.get_tensor_mode(name) == trt.TensorIOMode.INPUT:
-                self.input_shapes[name] = shape
-            else:
-                self.output_shapes[name] = shape
-
-        self.state = {}
-        for base_name in state_names:
-            name = self.resolve_name(base_name)
-            if name in self.input_shapes:
-                tensor = aligned_tensor(self.input_shapes[name], self.tensor_dtypes[name], "cuda")
-                tensor.zero_()
-                self.state[name] = tensor
-                self.context.set_tensor_address(name, tensor.data_ptr())
-        if self.state:
-            torch.cuda.synchronize()
-
-    def resolve_name(self, base_name):
-        if base_name in self.tensor_names:
-            return base_name
-        suffixed_name = f"{base_name}.1"
-        return suffixed_name if suffixed_name in self.tensor_names else base_name
-
-    def reset_state(self):
-        for tensor in self.state.values():
-            tensor.zero_()
-
-    def prepare_input(self, name, inputs):
-        shape = self.input_shapes[name]
-        base_name = name.rsplit(".1", maxsplit=1)[0] if name.endswith(".1") else name
-        if base_name not in inputs:
-            raise KeyError(f"Missing TensorRT input {base_name}")
-        value = inputs[base_name].to(device="cuda", dtype=self.tensor_dtypes[name])
-        if tuple(value.shape) != shape:
-            if tuple(value.shape[1:]) == shape:
-                value = value.squeeze(0)
-            elif tuple(shape[1:]) == tuple(value.shape):
-                value = value.unsqueeze(0)
-            else:
-                raise ValueError(
-                    f"Input {base_name} has shape {tuple(value.shape)}, expected {shape}"
-                )
-        return value
-
-    def __call__(self, stream, **inputs):
-        input_buffers = {}
-        for name, shape in self.input_shapes.items():
-            if name in self.state:
-                continue
-            value = self.prepare_input(name, inputs)
-            buffer = aligned_tensor(shape, value.dtype, value.device)
-            buffer.copy_(value)
-            input_buffers[name] = buffer
-            self.context.set_tensor_address(name, buffer.data_ptr())
-
-        outputs = {}
-        for name, shape in self.output_shapes.items():
-            output = aligned_tensor(shape, self.tensor_dtypes[name], "cuda")
-            outputs[name] = output
-            self.context.set_tensor_address(name, output.data_ptr())
-
-        if not self.context.execute_async_v3(stream.cuda_stream):
-            raise RuntimeError("TensorRT execution failed")
-        stream.synchronize()
-        return outputs
-
+from examples.onnx_ptq.trt_runner import TensorRTRunner
 
 STATE_NAMES = (
     "memory_embedding",
@@ -154,8 +49,7 @@ STATE_NAMES = (
 
 class Far3DDecoderRunner(TensorRTRunner):
     def __init__(self, engine_path, input_callback=None):
-        super().__init__(engine_path, STATE_NAMES)
-        self.input_callback = input_callback
+        super().__init__(engine_path, STATE_NAMES, input_callback)
         self.scene_token = None
         self.timestamp_offset = None
 
@@ -175,16 +69,6 @@ class Far3DDecoderRunner(TensorRTRunner):
                 device="cuda",
             )
         inputs["timestamp"] = (timestamp - self.timestamp_offset).float()
-        if self.input_callback:
-            calibration_inputs = {}
-            for name in self.input_shapes:
-                base_name = name.rsplit(".1", maxsplit=1)[0] if name.endswith(".1") else name
-                if name in self.state:
-                    value = self.state[name]
-                else:
-                    value = self.prepare_input(name, inputs)
-                calibration_inputs[base_name] = value
-            self.input_callback(calibration_inputs)
         outputs = super().__call__(stream, **inputs)
         for base_name in STATE_NAMES:
             input_name = self.resolve_name(base_name)
@@ -287,7 +171,7 @@ def main():
                 }
             }
         )
-        if args.max_samples is not None and len(outputs) == args.max_samples:
+        if args.max_samples is not None and len(outputs) >= args.max_samples:
             break
 
     if len(outputs) < len(dataset):
