+++ examples/onnx_ptq/petr/Dockerfile
@@ -1,5 +1,17 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 FROM nvcr.io/nvidia/pytorch:26.07-py3
 
@@ -12,9 +24,9 @@ RUN apt-get update && DEBIAN_FRONTEND=noninteractive apt-get install -y --no-ins
 
 ENV UV_PYTHON_INSTALL_DIR=/opt/python
 
-RUN python -m pip install --no-cache-dir uv && \
-    uv python install 3.8 && \
-    uv venv --seed --python 3.8 /opt/petr
+RUN python -m pip install --no-cache-dir uv==0.12.3 && \
+    uv python install 3.8.20 && \
+    uv venv --seed --python 3.8.20 /opt/petr
 
 COPY examples/onnx_ptq/petr/requirements*.txt /tmp/petr-requirements/
 RUN env -u PIP_CONSTRAINT /opt/petr/bin/python -m pip install --no-cache-dir \
