+++ examples/onnx_ptq/petr/requirements.txt
@@ -1,25 +1,25 @@
 --extra-index-url https://pypi.nvidia.com
 --find-links https://download.openmmlab.com/mmcv/dist/cu117/torch1.13.0/index.html
 
-einops
-ipython<9
-lyft-dataset-sdk
+einops==0.8.1
+ipython==8.12.3
+lyft-dataset-sdk==0.0.8
 mmcv-full==1.7.0
 mmdet==2.28.2
 mmsegmentation==0.30.0
-networkx>=2.2
-numba==0.57.1
-numpy<1.24
-nuscenes-devkit
-onnx
+networkx==2.2
+numba==0.53.0
+numpy==1.23.5
+nuscenes-devkit==1.1.11
+onnx==1.17.0
 onnx-graphsurgeon==0.6.1
-onnxruntime
-onnxsim
-opencv-python==4.5.5.64
-plyfile
-pycuda
-scikit-image
-setuptools<81
+onnxruntime==1.19.2
+onnxsim==0.5.0
+opencv-python==4.8.1.78
+plyfile==1.0.3
+pycuda==2026.1
+scikit-image==0.19.3
+setuptools==75.3.4
 tensorrt-cu13-bindings==11.1.0.106
-trimesh
+trimesh==2.35.39
 yapf==0.32.0
